Refactor code for manager dashboard and performance.js file
