+++ static/manager-dashboard/.gitignore
