+++ static/manager-dashboard/package.json
