+++ static/manager-dashboard/public/index.html
