+++ static/manager-dashboard/src/App.js
@@ -1,0 +1,14 @@
+import PerformancePieChart from './Components/PerformancePieChart';
+import PerformanceSatisfactionBarChart from './Components/PerformanceSatisfactionBarChart';
+
+function App() {
+    return (
+        <div>
+            <PerformancePieChart/>
+            <PerformanceSatisfactionBarChart/>
+        </div>
+
+    );
+}
+
+export default App;
