+++ static/manager-dashboard/src/index.js
