+++ static/manager-dashboard/src/shared/Chart.css
