+++ static/manager-dashboard/src/shared/Chart.js
