+++ static/manager-dashboard/src/shared/PieChart.js
