+++ static/manager-dashboard/src/shared/VerticalBarChart.js
